requirements.txt
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
django-debug-toolbar==2.2
djangorestframework==3.9.2
gunicorn==19.9.0
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
psycopg2-binary
mysqlclient==1.4.6